+++ src/ByteDev.Collections/EnumerableStringExtensions.cs
@@ -10,12 +10,12 @@ namespace ByteDev.Collections
     public static class EnumerableStringExtensions
     {
         /// <summary>
-        /// Get the longest string in <paramref name="source" />; otherwise null if <paramref name="source" /> is empty.
+        /// Returns the first longest string in <paramref name="source" />; otherwise null if <paramref name="source" /> is empty.
         /// </summary>
         /// <param name="source"><see cref="T:System.Collections.Generic.IEnumerable`1" /> to search for the longest string in.</param>
         /// <returns>Longest string in <paramref name="source" />.</returns>
         /// <exception cref="T:System.ArgumentNullException"><paramref name="source" /> is null.</exception>
-        public static string GetLongest(this IEnumerable<string> source)
+        public static string FirstLongest(this IEnumerable<string> source)
         {
             if (source == null)
                 throw new ArgumentNullException(nameof(source));
@@ -35,12 +35,12 @@ namespace ByteDev.Collections
         }
         
         /// <summary>
-        /// Get the shortest string in <paramref name="source" />; otherwise null if <paramref name="source" /> is empty.
+        /// Returns the first shortest string in <paramref name="source" />; otherwise null if <paramref name="source" /> is empty.
         /// </summary>
         /// <param name="source"><see cref="T:System.Collections.Generic.IEnumerable`1" /> to search for the shortest string in.</param>
         /// <returns>Shortest string in <paramref name="source" />.</returns>
         /// <exception cref="T:System.ArgumentNullException"><paramref name="source" /> is null.</exception>
-        public static string GetShortest(this IEnumerable<string> source)
+        public static string FirstShortest(this IEnumerable<string> source)
         {
             if (source == null)
                 throw new ArgumentNullException(nameof(source));